+++ patternsleuth/Cargo.toml
@@ -22,10 +22,10 @@ inventory = "0.3.14"
 itertools.workspace = true
 serde = { workspace = true, optional = true, features = ["derive"] }
 typetag = { version = "0.2.15", optional = true }
+gimli = "0.28.1"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 libc = { version = "0.2.152", optional = true }
-gimli = "0.28.1"
 
 [target.'cfg(windows)'.dependencies]
 windows = { workspace = true, optional = true, features = [
